# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cloud SaaS POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory levels
      parameters:
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: productId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/transactions:
    get:
      operationId: listInventoryTransactions
      tags: [inventory]
      summary: List inventory transactions/ledger
      parameters:
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: productId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Inventory transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"
    post:
      operationId: createInventoryTransaction
      tags: [inventory]
      summary: Record inventory transaction (stock in/out/adjust/transfer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryTransaction"

  # ─────── CUSTOMERS ───────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /expense-categories:
    get:
      operationId: listExpenseCategories
      tags: [expenses]
      summary: List expense categories
      responses:
        "200":
          description: Expense categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseCategory"
    post:
      operationId: createExpenseCategory
      tags: [expenses]
      summary: Create expense category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseCategory"

  # ─────── REPORTS ───────
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [productId, productName, branchId, branchName, quantity, updatedAt]
      properties:
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: ["string", "null"]
        branchId:
          type: integer
        branchName:
          type: string
        quantity:
          type: number
        reorderLevel:
          type: number
        costPrice:
          type: number
        totalValue:
          type: number
        updatedAt:
          type: string
    InventoryTransaction:
      type: object
      required: [id, productId, branchId, type, quantityBefore, quantityChanged, quantityAfter, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        type:
          type: string
          enum: [stock_in, stock_out, adjustment, transfer_in, transfer_out, damaged, sale, purchase]
        quantityBefore:
          type: number
        quantityChanged:
          type: number
        quantityAfter:
          type: number
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    InventoryTransactionInput:
      type: object
      required: [productId, branchId, type, quantityChanged]
      properties:
        productId:
          type: integer
        branchId:
          type: integer
        type:
          type: string
          enum: [stock_in, stock_out, adjustment, transfer_in, transfer_out, damaged]
        quantityChanged:
          type: number
        toBranchId:
          type: ["integer", "null"]
        reference:
          type: string
        notes:
          type: string

    # ─────── CUSTOMER ───────
...
    Expense:
      type: object
      required: [id, amount, description, date, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        branchId:
          type: ["integer", "null"]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        reference:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [amount, description, date]
      properties:
        categoryId:
          type: ["integer", "null"]
        branchId:
          type: ["integer", "null"]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        reference:
          type: string
    ExpenseCategory:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
    ExpenseCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string

    # ─────── REPORTS ───────
...